pyproject.toml
[tool.poetry]
name = "generate-systemd-timer"
version = "0.1.0"
description = "Generate a systemd unit.timer and unit.service pair"
license = "MIT"
authors = [
    "Thom Wiggers <thom@thomwiggers.nl>",
]

packages = [
    { include = "systemd_generator" },
]

[tool.poetry.dependencies]
python = ">=3.6"
jinja2 = "^2.11"
python-editor = "^1.0"

[tool.poetry.scripts]
generate-systemd-timer = 'systemd_generator:main'
